requirements.txt
PyQt5
pygame
Pillow
groq
python-dotenv